# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Plataforma MOOC API
  version: 0.2.0
  description: >
    API REST de la plataforma web de cursos masivos abiertos en línea.
    Prefijo de versión: /api/v1. Errores uniformes ({"error":{"code","message","details"}}),
    sesiones por cookie httpOnly o cabecera Authorization: Bearer <token>.
    Este documento cubre los endpoints implementados en la iteración actual
    (identidad, administración, autoría de cursos, catálogo e inscripciones).
    Pendientes: quizzes, progreso e insignias (ver README).
servers:
  - url: /api/v1
security:
  - sessionCookie: []
  - bearerAuth: []
paths:
...
  /courses/versions/{versionId}/resources/{resourceId}:
    patch:
      summary: Actualiza un recurso
      operationId: updateResource
      parameters:
        - { name: versionId, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: resourceId, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ResourceInput" }
      responses:
        "200": { description: Recurso actualizado. }
    delete:
      summary: Elimina un recurso
      operationId: deleteResource
      parameters:
        - { name: versionId, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: resourceId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "204": { description: Recurso eliminado. }
...
  /badges/{code}/openbadge:
    get:
      tags: [Insignias]
      summary: La insignia como credencial Open Badges 3.0
      description: >
        Open Badges 3.0 deja de ser una imagen con metadatos y pasa a ser una
        Credencial Verificable del W3C. Esta respuesta es el documento JSON-LD
        sin firmar, legible y apto para inspección.

        No lleva nada de la persona: el sujeto se identifica por la propia URL
        de verificación. Una insignia revocada responde 409: una credencial
        firmada no se puede desdecir, y para el estado está la URL pública.
      operationId: openBadgeCredential
      security: []
      parameters:
        - { name: code, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: Credencial JSON-LD.
          content:
            application/ld+json:
              schema: { type: object }
        "409": { description: La insignia fue revocada. }
        "404": { $ref: "#/components/responses/Problema" }
  /badges/{code}/openbadge.jwt:
    get:
      tags: [Insignias]
      summary: La credencial firmada como VC-JWT
      description: >
        La credencial como carga de un JWS firmado con Ed25519. Se elige esta
        prueba y no una Data Integrity porque las pruebas de integridad exigen
        canonicalizar el JSON-LD, y una canonicalización a medias produce
        firmas que unos verificadores aceptan y otros no, que es peor que no
        firmar.

        Se verifica con la clave publicada en /.well-known/jwks.json, sin
        volver a llamar a esta plataforma. Un despliegue sin clave configurada
        responde 503: las insignias siguen siendo verificables por su URL
        pública, y lo único que falta es la credencial portátil.
      operationId: openBadgeJWT
      security: []
      parameters:
        - { name: code, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: VC-JWT.
          content:
            application/vc+ld+json+jwt:
              schema: { type: string }
        "503": { description: El despliegue no tiene clave de firma configurada. }
...
  /courses/versions/{versionId}/resources/{resourceId}/revisions:
    get:
      tags: [Autoría]
      summary: Historial de revisiones de un recurso
      description: >
        No incluye el contenido de cada revisión: un historial largo de un
        recurso extenso pesaría megabytes, y la lista solo necesita decir quién
        guardó y cuándo. El contenido se pide por revisión.
      operationId: listRevisions
      parameters:
        - { name: versionId, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: resourceId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200":
          description: Revisiones, la más reciente primero.
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/Revision" }
    post:
      tags: [Autoría]
      summary: Guarda el contenido y deja una revisión
      description: >
        Actualiza el contenido vigente del recurso y añade una revisión. Si el
        contenido es idéntico al de la última responde 200 con
        `status: sin_cambios` en lugar de crear una: el autoguardado dispara
        cada pocos segundos y, sin esa guarda, media hora de escritura dejaría
        cientos de revisiones idénticas.
      operationId: saveRevision
      parameters:
        - { name: versionId, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: resourceId, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [content_md]
              properties:
                content_md: { type: string }
      responses:
        "201":
          description: Revisión creada.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Revision" }
        "200": { description: El contenido no cambió; no se creó ninguna revisión. }
  /courses/versions/{versionId}/resources/{resourceId}/revisions/{numero}:
    get:
      tags: [Autoría]
      summary: Una revisión concreta, con su contenido
      operationId: getRevision
      parameters:
        - { name: versionId, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: resourceId, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: numero, in: path, required: true, schema: { type: integer, minimum: 1 } }
      responses:
        "200":
          description: Revisión con su contenido.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Revision" }
        "404": { $ref: "#/components/responses/Problema" }
  /courses/versions/{versionId}/resources/{resourceId}/revisions/{numero}/restore:
    post:
      tags: [Autoría]
      summary: Restaura el contenido de una revisión anterior
      description: >
        No borra ni retrocede el historial: la restauración se guarda como una
        revisión nueva. Así queda constancia de que alguien restauró, y deshacer
        una restauración equivocada es otra restauración, no una pérdida.
      operationId: restoreRevision
      parameters:
        - { name: versionId, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: resourceId, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: numero, in: path, required: true, schema: { type: integer, minimum: 1 } }
      responses:
        "200": { description: Contenido restaurado, con la revisión que lo registra. }
        "404": { $ref: "#/components/responses/Problema" }
...
components:
...
  responses:
...
    Problema:
      description: >
        Error uniforme. El campo `error.code` identifica el motivo de forma
        estable; `error.details` enumera los motivos cuando hay varios, como en
        la validación de publicación.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: mooc_session
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    Revision:
      type: object
      required: [id, resource_id, revision_number, created_at]
      properties:
        id: { type: string, format: uuid }
        resource_id: { type: string, format: uuid }
        revision_number: { type: integer }
        content_md:
          type: string
          description: Solo al pedir una revisión concreta, no en la lista.
        author_email: { type: string }
        created_at: { type: string, format: date-time }
...
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string }
            message: { type: string }
            details: { type: array, items: { type: string } }
...
    ResourceInput:
      type: object
      required: [type, title, position]
      properties:
        type:
          type: string
          enum: [text, image, video, audio, pdf, presentation, file, iframe, link, quiz]
        title: { type: string }
        position: { type: integer }
        visible: { type: boolean }
        required: { type: boolean }
        downloadable: { type: boolean }
        text_content_md: { type: string }
        external_url: { type: string }